update mycomponentclass.js and mycomponent.js
